compile time fix
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@3495 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/source_mod.f90
@@ -46,7 +46,7 @@ module source_mod
     subroutine writeSourceList(source, nSource)
       type(SOURCETYPE) :: source(:)
       integer :: nSource
-      real :: fromSpec
+      real(double) :: fromSpec
       real(double) :: tot
       integer :: i, j
